Add fleeing behaviour for units that occupy buildings.
This adds a monitor that detects when the enemy unit count is below some
level.  When this happens, units that are occupying buildings from the
buildingOccupy function flee the AO.

This change tries to improve the situation when enemy units must be
cleared and only a couple are left hiding in buildings.  Rather than
forcing players to search an urban area for 1 or 2 units, the units just
flee.