Fix respawn bug where players doesn't get proper loadout. (#52)
The bug was caused by a regression when the mission events moved
to the server.  The den_insert event was used locally to detect when a
player was at the staging area.  This broke when den_insert was no
longer global.  The fix is to replace this logic with a local trigger
that detects when a player is outside the staging area.