Improve MP respawn.
This change improves MP repawn by removing the respawn menu with respawn
in place.  Respawning from a static position like the start LZ is
usually not ideal from a game play perspective for a couple of reasons:
1) in some cases it allows players to teleport to the LZ if gettin back
to the LZ is a task, 2) it can force player to start over during a
mission.