Add support for mission location selection. (#75)
Players can multi-select location on the map during the breifing.  If
more than one location is selected, a random location from that set is
selected.  If a mission that a player selected cannot be initialized,
then another location is chosen as a default.

This change also updates the location blacklist for each terrain to
improve the performance of location selection.  The fucntion
fn_locationBlacklist can be used to automate if a location should be
blacklisted.