Add RHSPKL terrain. (#72)
In addition to RHSPKL terrain, this change adds support for custom
locations on a terrain.  This allows for better utilization of the
terrain.

Custom locations can  be added to the map by calling fn_createLocation.
This just adds a position and name to a global array.  This array is
included in the location list prodcued by fn_locations