Improve the spawn logic for Defend Convoy mission (#110).
Enemy groups are spawned if either the kill threshold is met, or
enough time has elapsed. The elapsed time condition is to prevent
the mission from getting stuck if enemies are hiding.

If the convoy position is seized by the enemy, the mission ends
if failure. This is to prevent players from hiding out waiting
for time to expire.